Add N parameter to the grass cover erosion inwards, outwards and height structures failuremechanism database entities WTI-1531

+++ Application/Ringtoets/src/Application.Ringtoets.Storage/Create/GrassCoverErosionInwards/GrassCoverErosionInwardsFailureMechanismCreateExtensions.cs
@@ -54,7 +54,7 @@ namespace Application.Ringtoets.Storage.Create.GrassCoverErosionInwards
         {
             var metaEntity = new GrassCoverErosionInwardsFailureMechanismMetaEntity
             {
-                N = (int) mechanism.GeneralInput.N,
+                N = mechanism.GeneralInput.N,
                 DikeProfileCollectionSourcePath = mechanism.DikeProfiles.SourcePath
             };
             entity.GrassCoverErosionInwardsFailureMechanismMetaEntities.Add(metaEntity);
